Android Spinner Example
Android Spinner is like the combox box of AWT or Swing. It can be used to display the multiple options to the user in which only one item can be selected by the user.
Android spinner is like the drop down menu with multiple values from which the end user can select only one value.
Android spinner is associated with AdapterView. So you need to use one of the adapter classes with spinner.